pyproject.toml
[tool.poetry]
name = "pictures-project"
version = "0.1.0"
description = ""
authors = ["juliabek <julia.bekish@innowise.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
djangorestframework = "^3.14.0"
celery = "^5.3.6"
python-dotenv = "^1.0.1"
django-celery-results = "^2.5.1"
psycopg2-binary = "^2.9.9"
redis = "^5.0.3"
django-celery-beat = "^2.6.0"
requests = "^2.31.0"
boto3 = "^1.34.63"
openai = "^1.14.1"
fuzzywuzzy = "^0.18.0"
levenshtein = "^0.25.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
